Add functions for finding where probes cover a sequence
For a given sequence and list of probes, these find for each probe
the portions (i.e., list of intervals) of the sequence that are
"covered" by the probe, where the test of whether a probe covers
a portion of a sequence is determined by a given function. For
example, a probe P could cover an interval I of the sequence if
P has a substring (with at most some number of mismatches) of
some specified length in common with I. This search is done in a
probabilistic setting (i.e., there is a potential for missing
intervals covered by a probe) by looking up kmers in the sequence
against a pre-constructed dict mapping kmers to probes.

The function probe.find_probe_covers_in_sequence is missing tests
and the added fuctions do not yet contain descriptive comments.