Increase number of kmers chosen in one adapter_filter test
This test would fail occassionally because of randomness involved
in adapter_filter (specifically, its search through a sequence of
which probes cover which portions of the sequence). Increasing the
number of chosen kmers greatly reduces the probability that this
test fails, without the need to seed the random number generator.