Allow targeted genomes with multiple chromosomes
This is needed for some pathogens, like malaria. The dataset
file should define whether the genomes in that dataset are made
up of more than one chromosome. If so, the dataset file should
store the paths to one or more FASTA files (returned by a
fasta_paths() function) where each FASTA file corresponds to one
genome/sample and gives the sequences for each chrosomsome in that
genome.

To make some code here (and hopefully in the future) slightly easier
to work with, this introduces a genome module with a Genome class.
Previously a genome was just made up of one sequence, so a simple
string sufficed; however, with more than one sequence in a genome,
this could prove useful.