Add identification option
When "--identify" is set, set_cover_filter chooses candidate
probes that make it easy to distinguish groupings (e.g., species)
of the target genomes in a sample. It tries to choose candidate
probes that only 'hit' one grouping, and only chooses probes that
hit more than one grouping if absolutely necessary to achieve the
required coverage.

Also, this commit modifies the technique for blacklisting genomes
in set_cover_filter to use a similar approach as the identification
method. That is, it uses 'ranks' rather than 'costs' to penalize
candidate probes that cover portions of blacklisted genomes.