Run autopep8 on all Python code
autopep8 was run with aggressiveness level 2 ("-a -a"), which makes
non-whitespace changes as well as whitespace changes. autopep8 also
runs lib2to3 to make this code more compatible with Python 3. The
bulk of changes results from changing the indentation from 2 spaces
to 4 spaces; this causes the large diff (a change on almost every
line of code).