Revert (manually) some yapf changes
yapf (which is in alpha stage) made some changes that hurt readability;
some of these did violate the PEP 8 style guide and some did not. This
commit manually reverts some changes made by yapf in order to be more
compliant with PEP 8 and improve readability.

As of this commit, there are just six PEP 8 violations. Each of these is
"line too long" (the longest line of these six is 85 chars; the max
allowed is 79). Fixing these would hurt readability, so these six PEP 8
violations were not fixed.