Optimize invalidation of intersects in set cover
When using interval sets, there is no need to invalidate the
intersection counts between a modified universe and all sets.
Instead, it is possible to determine which sets' intersection
counts with a modified universe might change (and therefore
should be invalidated) and which cannot possibly change (and
therefore should not be invalidated). By invalidating fewer
memoized values compared to before, this provides a substantial
improvement in runtime.