Select k-mers by "pigeonholing" mismatches
When the number of allowed mismatches is small compared to the
length of the probes and the desired alignment (lcf_thres) is
the full probe, it is possible to build a map from k-mers to
probes in a way that is more space efficient and provides more
time efficiency when using the map in probe.find_probes_covers_in_sequence.
Namely, assuming the probe length (e.g., 100 bp) has many factors,
we can pick k-mer length k such that enough non-overlapping k-mers
can be extracted from the probe to guarantee that we find a match
with a sequence that is within the given number of mismatches from
the probe (i.e., there is some k-mer in that sequence that matches
exactly one of the extracted k-mers from the probe). When this is
possible, it offers the following advantages compared to the approach
of randomly selecting k-mers from probes:
 - Fewer k-mers can be chosen from each probe, meaning the map from
   k-mers to probes is smaller.
 - Since fewer k-mers are taken, fewer false positives arise in
   probe.find_probe_covers_in_sequence (i.e., fewer k-mer matches
   are found that do not yield true alignments), so it runs
   faster.
 - It guarantees that if a probe aligns to a sequence to within the
   given number of mismatches, then
   probe.find_probe_covers_in_sequence will find the alignment using
   this k-mer to probe map. With the approach of randomly selecting
   k-mers from probes, this is not necessarily true because all
   the selected k-mers may happen to be ones that mismatch with
   the sequence to which the probe aligns.
When it is not possible to build the map from k-mers to probes with
the pigeonholing approach, the call to
probe.construct_kmer_probe_map_to_find_probe_covers will instead
resort to the (old) approach of building the map by randomly
selecting k-mers in probes.