Add 'cover_groupings_separately' option
When enabled, a separate set cover instance is created and solved
for each grouping of target genomes. The results (probes) are then
pooled together, and their union composes the selected probes.
This improves runtime but will generally result in a larger number
of probes than when the option is not enabled, particularly when
target genomes across groupings are similar at a nucleotide level.