Improve memory usage for input with many genomes
When there are many candidate probes across many genomes, the memory
usage could be substantial. The primary culprit is the sets data
structure, which is created by filter.set_cover_filter for input to
the utils.set_cover.approx_multiuniverse function. Consider some
candidate probe that is given set id S. This creates a dict sets[S],
which maps universes to intervals that give the regions of each
universe that are covered by the probe. If the candidate probe covers
part of a universe (genome) U, then consider sets[S][U] to be an entry
in the data structure. Note that, in practice, there can be hundreds
of millions of entries; for example, in the hepatitis_c dataset there
are roughly 170,000 candidate probes across roughly 1,130 genomes,
which can yield over 192 million entries (this is in the worst-case,
where every candidate probe covers a part of every genome).
Previously, this entry would be stored as an instance of
utils.interval.IntervalSet. An instance of a class carries substantial
overhead, which is very noticeable when there are so many of them. But
note that this entry is typically just a single interval because most
probes cover at most one interval of a universe/genome. The changes
seek to exploit this fact in order to reduce memory requirements.
Below are the primary changes.

In hybseldesign.filter.set_cover_filter:
- When building the sets structure, now adds an interval (as a tuple)
  directly rather than putting the interval inside a list. This is
  useful because a list has a lot of overhead and most probes align to
  just one interval in a universe/genome.
- When finalizing the sets structure, now only converts the intervals
  of a set to an interval.IntervalSet if there are more than one
  interval in the set. If there is just one, the interval continues to
  be stored directly as a tuple.

In hybseldesign.utils.interval:
- Converts IntervalSet to a new-style class and defines a __slots__
  field. An IntervalSet may be instantiated many times; indeed,
  without the above change (storing a single interval directly), in
  practice it could be instantiated hundreds of millions of times.
  Python normally creates a dict for each instance of a class, which
  adds a huge amount of overhead. The __slots__ field prevents Python
  from creating this dict, at the cost of a few limitations on how the
  class can be used.

In hybseldesign.utils.set_cover:
- If use_intervalsets is True and a set (contained within the sets
  argument) contains just a single interval, allows the interval to be
  supplied directly as a tuple rather than as an interval.IntervalSet.
  The approx_multiuniverse function will automatically convert the
  single interval to an instance of interval.IntervalSet as needed.