Improve running time of set cover implementation
In practice (at least in the application to hybrid selection),
often the minimum ratio for a set computed in one iteration of
the greedy set cover algorithm is equal to the minimum ratio
from the previous iteration. This commit has changes to exploit
this fact in order to speed up the implementation. In particular,
when this is the case on an iteration, that iteration can
take just a tiny fraction of the time it would have taken
without the changes in this commit. When this is not the
case on an iteration, that iteration will still take the same
amount of time even with these changes.