Add 'cover_extension' parameter
Add a parameter that allows a probe to cover (or "capture") some
number of bp on each side of the sequence to which it hybridizes.
In particular, a probe is said to cover the region to which it
hybridizes as well as 'cover_extension' bp on each side of that
region. This more realistically models hybrid selection because
a probe hybridizes to a fragment of DNA that includes both the
intended sequence target of the probe, as well as the region
surrounding and including that targeted sequence. Raising this
parameter can significantly reduce the number of probes required
to achieve a desired coverage.