Add option to filter probes using a FASTA file
This option is useful for processing probes after they have already
been designed using the set cover filter. For example, if they have
already been designed and are saved to a FASTA file, we can supply
that file as an argument (--filter_from_fasta) and also include
--skip_set_cover to avoid re-running the set cover filter; then,
we can add adapters to the probes (if they had not been added) and
also perform a coverage analysis.