Allow for overly-large specified coverage (in bp)
Previously, an exception would result if the desired coverage
(in bp) is greater than the size of a genome. This change now
allows for such an overly-large desired coverage by adjusting
the coverage, if it is larger than the genome size, to be
equal to the size of the genome (i.e., to achieve full genome
coverage).