Uses multiple processes to find probe covers
A pool of processes is created (using the multiprocessing module) based
on a set of probes P where we wish to find regions in sequences that are
covered by P. Opening this pool is accompanied by the creation of a
k-mer map (i.e., map from k-mers to probes) that can be shared in memory
across processes. Then, a sequence S is searched using this pool; S is
subdivided into regions and each worker (process) is assigned one of the
regions of S to search. Multiple sequences can be searched using a pool;
the sequences are searched serially (i.e., only one sequence at a time)
but each sequence itself is searched in parallel (i.e., multiple regions
of the sequence at once). At the end, the pool should be closed.

The SharedKmerProbeMap takes a dict, of the particular form used by
the k-mer map, and converts it into a collection of arrays that can
be shared across processes. This avoids having to copy the full
k-mer probe map to each process; it may be large in memory, so copying
it to each would require N times as much memory for N processes
and may also take considerable time.

The multiprocessing module in Python 2.7 seems to be buggy, so much of
the code in hybseldesign/probe.py is designed to avoid these bugs.
For example, some calls (like Pool.join()) may hang indefinitely, so
this introduces timeouts around those calls. Without these fixes, the
bugs were rare -- running all unit tests 1000 times, only a few unit
tests would fail in total. But these fixes appear to avoid any bugs
from causing any unit tests to fail.

Occassionally a process (created in a pool) will fail to exit.
The job submitted to LSF that spawned this process may then fail to
exit, even if it (successfully) completes all of its work. It should be
safe to kill this job (e.g., with bkill). Similarly, if all unit tests are
submitted as a job to LSF (i.e., via 'python -m unittest discover'), the
job may occassionally never end even though all unit tests pass.