Change keys()/values() to iterkeys()/itervalues()
The changes are only made when appropriate (i.e., a list is not
needed). This helps prior to running 2to3 because that program
changes x.keys() to list(x.keys()) and changes x.iterkeys()
to x.keys(); in Python 3, keys() returns an iterable but in
Python 2 it gives a list.