Allow >1 FASTA file for single-chromosome datasets
When more than one FASTA file are given for a single-chromosome
dataset, they are effectively treated as if one FASTA file was
given with the concatenation of the multiple FASTA files. This
allows a dataset's sequences (genomes) to come from more than one
FASTA file, even if the distinction between the files has no
meaning.