Add option of providing a collection of datasets
A collection of datasets lists multiple datasets, so that
all of the ones in the collection can be used as input to
a run of the program. Effectively, a collection provides a
shortcut for specifying many datasets (all falling under
a shared category, like viruses with a human host) as
input without having to list all of them in the command
line arguments.