Add all sequenced viral genomes with human hosts
The genomes added encompass all sequences for viruses on
GenBank whose hosts include human. They are from the
"Accession list of all viral genomes" at
  http://www.ncbi.nlm.nih.gov/genome/viruses/
(taxid10239.nbr.txt). The sequences were merged with
existing sequence information for the viruses already present
in this repo -- i.e., for these viruses, any sequences that
were not a part of the GenBank list, such as some MERS
sequences from virological.org, were added to the
downloaded sequences. When possible, the sequences for
viruses whose genomes are segmented were automatically grouped
(each grouping being a genome) based on the sequences'
strain and/or isolate. The genomes were grouped into
'datasets' based on the virus' "Taxonomy name" given in
the accession list. A 'dataset' is usually a species
(sometimes with a shortened name), but may sometimes be a
genus. Each 'dataset' is given a Python file and the header
comment of this file gives a regex pattern that matches
the taxonomy name for all the genomes belonging to the
dataset.

This commit also includes a 'collection'
  (hybseldesign/datasets/collections/viruses_with_human_host.py)
that contains all the datasets with sequences whose hosts
include human.

Some already existing datasets were removed and their
sequences are included in newly named datasets. For example
the 'ebola2014' dataset is removed; its sequences are
now a part of the 'ebola_zaire_with_2014' dataset. Because
some test cases used the ebola2014 dataset, these test cases
have been modified to instead use the ebola_zaire_with_2014
dataset.