Breakup long test case
Travis CI sometimes gives an error on the
  hybseldesign.tests.test_probe.TestFindProbeCoversInSequence.test_random_long_genome
test case because it takes too long to run (> 10 mins). This
breaks its two iterations into two separate test cases. It also
sets n_workers, for the probe finding pool, at 2, to avoid
leaving the number of workers dependent on the host.