Add HIV (1+2) datasets in which LTRs are stripped
Using the sequences in the 'hiv1' and 'hiv2' datasets,
this adds the corresponding 'hiv1_without_ltr' and
'hiv2_without_ltr' datasets in which the LTRs are
stripped from both ends of each genome.

This also updates the 'viruses_with_human_host' collection
to change the HIV datasets to be those without the LTRs --
i.e., to contain 'hiv1_without_ltr' and 'hiv2_without_ltr'
instead of 'hiv1' and 'hiv2'.