Improve runtime of set_cover.approx_multiuniverse
This change results in fewer invalidations of memoized
intersection counts, which should improve runtime
in practice. In particular, it is useful for (long)
genomes in which a single candidate probe may map to a
region at the start of the genome and a region at the
end of the genome; this change results in fewer
invalidations of the intersection counts for these
kinds of candidate probes.