Increase time limits when opening/closing pools
For many applications, the time limits on opening/closing
probe finding multiprocessing pools were too small. In
particular, this was the case when opening a pool after
another pool had earlier been opened and closed; it would
often take longer than 5 secs, meaning it would take a
very long time (or infinite) to open a pool because it
would have to keep retrying until it could be opened
within 5 secs. Increasing the time limit for opening a
pool to 60 seconds (and then growing the limit
exponentially on successive tries) alleviates this issue.