Fix bug in utils.set_cover.approx_multiuniverse
The list rank_vals only needs to store distinct rank values;
prior to this commit, it was storing all rank values (many
are repeated, so the size of rank_vals could be as large as
the number of probes). It was not incorrect per se -- the
function would still terminate and have the correct output --
but it meant the loop would sometimes go through many more
iterations than needed, where in each iteration curr_rank_index
would be incremented but nothing else would change, and
therefore the function would take far longer than needed.