Add 14 Zika virus sequences
Previously there were 2 Zika virus sequences (both complete
genomes). This adds 2 more complete genomes: one (NC_012532)
is the Zika reference sequence and the other (KU321639) was
added to GenBank on December 28, 2015. The other 12 new
sequences each cover the full coding sequence ('complete cds'),
which is most of the Zika genome. This brings the total number
of Zika sequences to 16.