Add 'island_of_exact_match' option
This option provides the ability to require that a
probe can hybridize to a region of a sequence only
if (among the other conditions) there is a stretch
(island) of 100% identity (no mismatches) that has
length 'island_of_exact_match' bp between the probe
and the target sequence. When 'island_of_exact_match'
is 0 (default value), this is equivalent to not
having the requirement.

This was added after a comment from NimbleGen that
said it is bad (in terms of hybridization) when
mismatches are distributed evenly through a probe.
Rather, it is important that there are "islands"
or stretches of 100% identity between the probe and
the target sequence. They listed 30 bp as a
conservative estimate for the length that these
islands should be (so island_of_exact_match = 30)
to ensure the probes will capture well. The value
probably does not have to be this high.