Add 'skip_reverse_complements' option
This option tells the bin/make_probes.py script not
to add to the output the reverse complement of each
probe. (The default behavior is to add the reverse
complements.) Furthermore, when set, the analysis
performed of the probes is not done on the reverse
complements of each target genome.