Add '--expand_n' option
A probe may contain one or more 'N' bases, as determined
by the input target sequences. If this is the case,
sometimes we are able to synthesize these probes so that
the 'N' bases are replaced by all four real bases (i.e.,
the probe 'ANA' becomes synthesized as the four probes
'AAA', 'ATA', 'ACA', 'AGA').

However, if it is not possible to do such a synthesis,
we can effectively have the probe design software do
such an expansion. The '--expand_n' option does this:
the probes that contain 'N' are replaced by additional
probes in which the 'N' bases are substituted with real
bases ('A', 'T', 'C', and 'G'). This is done
combinatorially when probes contain more than one 'N',
so that if a probe has n 'N' bases, it is replaced with
4^n new probes; therefore, this should only be used when
few probes contain 'N' bases and/or when the number of 'N'
bases in each probe is small.