Add seven datasets
The added datasets are:
 - 'mastadenovirus_e'
 - 'mastadenovirus_f'
 - 'mastadenovirus_g'
 - 'parechovirus_a'
 - 'parechovirus_b'
 - 'semliki_forest'
 - 'sindbis'

All of these datasets infect humans or have reported cases of
human infections. They are all added to the list of
viruses_with_human_host.

The sequences in these datasets all come from the NCBI's viral
accession list. These were previously not included as datasets
because the NCBI's viral accession list (mistakenly) does not
list human as a host for these sequences.