Remove duplicate sequences from datasets
There was a bug in the script that automatically downloads
sequences from NCBI and groups them into datasets. When
given a list of existing sequences to merge with, it would
sometimes create duplicate sequences. If an existing sequence
had accession number N with a version, yet the accession list
used for downloads has N without a version, then both the
existing sequence and the downloaded sequence would be added
to a dataset even though these are the same. (For example, if
an existing sequence is AF123.1 and the accession list used
for downloads has AF123, then both AF123.1 and AF123 were
added as sequences to a dataset.)

There is one other rare case in which the same sequence might
be added twice to a dataset. If the header of a sequence
from NCBI differs from the header of an existing sequence
being merged with the downloaded sequences -- yet these two
sequences are the same -- then they would have been added
twice.

A commit (in a separate repository) fixes the issue in the
download script so that sequences are not added twice. This
commit adds the resulting datasets (i.e., the ones without
duplicate sequences).