Add four datasets
The added datasets are:
 - 'bourbon'
 - 'cowpox'
 - 'monkeypox'
 - 'spondweni'

All of these datasets infect humans or have reported cases of
human infections. They are all added to the list of
viruses_with_human_host.

'bourbon' is a recently discovered virus and does not have
a reference genome on NCBI. Therefore, it is not on NCBI's
viral accession list.

'cowpox' and 'monkeypox' sequences come from the NCBI's viral
accession list. These sequences were not previously included
as datasets because the NCBI's viral accession list (mistakenly)
does not list human as a host for these.

'spondweni' sequences were sequenced very recently and were
only recently added to the NCBI's viral accession list.