Add Influenza A/B/C sequences
Prior to this commit, only some segments of the RefSeq genomes were
included in the influenza_{a,b,c} datasets. This is because NCBI
distributes Influenza accessions separately from their viral accession
list; when the last update to those datasets were made from the
accession list, it only included the few Influenza sequences that are
part of the accession list. (Prior to the last update, these datasets
included additional Influenza sequences because they included
sequences that were downloaded separately from a download based on the
accession list.)

This commit uses NCBI's Influenza Virus Resource to find accessions of
Influenza sequences for download. It includes all full-length
Influenza sequences that have human as a host.

Because there are so many sequences for Influenza A and Influenza B,
this puts all sequences across all genomes into a single file (rather
than creating a separate file for each genome) for each type and gzips
the files.