Improve runtime of determining whether two probes are redundant by LCS
This helps improve the runtime of the dominating set filter when
it determines whether two probes are redundant based on a longest
common substring up to k mismatches. Namely, it anchors the
determination of the longest common substring on a shared
k-mer between two probes. It is faster to compute the longest
common substring given this anchor than it is without. We already
have the anchor if we use a heuristic to check whether two probes
might be redundant (i.e., probe.shares_some_kmers(..)).