Add tools to pool probes across datasets while optimizing parameters
This adds param_search.py, which contains functions for searching the
space of parameters across datasets. It finds the optimal parameters
for each dataset, according to a loss function, subject to a
constraint on the total number of probes pooled across the datasets.
Currently, it optimizes the mismatches and cover_extension parameters.
This is heavily based on the script find_optimal_params.py in
@haydenm's hybsel-design-runs repo.

This also adds a script, pool_probes.py, that exposes these
functions.

Because this uses scipy's optimize package for minimizing the loss
function, this commit adds scipy as a requirement.

Additionally, this adds test functions. Some are integration tests
using data generated in the design of probe sets.