Add ability to search over more than 2 parameters
Previously, the parameter optimization procedure was specialized
to only the mismatches and cover_extension parameters (n=2). This
adds a higher dimensional search function that allows input
of n>2 parameters. To do so, it requires interpolating probe
counts over unstructured data in n>2 dimensions; it uses scipy's
interpolate package to do this.

The downside is that, unlike with the n=2 search, after optimizing
parameters this does not round them (e.g., to integral values
or to fit on a grid). This could be done, but is not implemented.

The 'standard' search (over only mismatches and cover_extension)
remains the default option.

There are integration tests for this N-D search, but it has
not been applied in practice on data.