Add coverage config file that omits files from coverage report
This adds a config for the coverage tool (used here as a plug-in to
nose). The config file is .coveragerc, which should be the default
that is read. On Travis, the coverage plug-in to nose will read this
config file when measuring code coverage, and the results should be
reflected in the upload to coveralls.