Exclude adapters by default
Previously, adapters were added to the end of each probe sequence
by default, but this could be left out using --skip-adapters.
This change makes it so that adapters are not added by default;
they can now be added to probe sequences using --add-adapters.

This also reorganizes/simplifies how the filters list is generated.