Add filter for near-duplicate probes based on Hamming distance
Like the duplicate filter, this is not necessary, but running it before the set
cover filter can significantly lower runtime and reduce memory usage by
decreasing the input size (number of candidate probes) to the set cover filter.
This detects and collapses near-duplicate probes using locality-sensitive
hashing with a family of hash functions that works with Hamming distance. An
option to use this filter (in place of the duplicate filter) is provided in the
arguments for design.py. There are two caveats explained in the argument's help
message: (1) this filter may lead to a slightly less-than-optimal solution
(i.e., more probes than needed) because it may remove "better" candidate probes
that are near-duplicates of ones that are kept, and (2) this filter may
(depending on the provided distance threshold) lead to less than desired
coverage of the target genomes because the universe to cover is constructed
from the set of candidate probes, which are collapsed by this filter.