Add filter for near-duplicate probes using MinHash
This filter is like the one using LSH based on Hamming distance, except this
uses a family that constructs MinHash signatures. This commit adds an option to
design.py for using this filter. The main advantage of this filter compared to
the Hamming distance filter is that it can be more sensitive in detecting
near-duplicate probes (e.g., ones that are shifted relative to each other). The
main downside is that the distance threshold is based on Jaccard distance and
it is less intuitive to understand how these values translate to probe
hybridization compared to Hamming distance.