Add error message explaining why probe count constraint cannot be satisfied
When the limit on the number of probes to design is not satisfied, this
now raises an custom exception (CannotSatisfyProbeCountConstraintError)
with a message explaining the likely reason. Previously, only an
assertion failed and it may have been unclear why. This also raises the
exception in the (n>2)-dimensional search, rather than only in the
standard search.