Expose option to set kmer_probe_map_k parameter
This parameter specifies the k-mer length used when constructing a map of
k-mers to probes.  (In particular, CATCH tries to determine a suitable k-mer
length using this value as a lower bound; if it cannot, it takes a randomized
approach and uses this value as the k-mer length.)

This is useful when designing short probes (e.g., meant to be primers), in
which case the default k-mer length may be too short relative to the probe to
find mappings of the probes to the target sequence.