Remake .py files for some segmented datasets
The sequence data for some datasets was stored as segmented genomes,
but the .py files treated them as being unsegmented. This was because
of a bug (now fixed) in how the dataset generation script determined
whether a genome is segmented. This adds the newly generated .py files
for these datasets, treating them as being segmented.

This fixes #17.