Add probe sets designed to target Influenza A, B, and C viruses
This adds two probe sets: one with 45,000 probes and another with
6,000 probes (twice these numbers after counting reverse complements,
which were added to the probe set). They are 100 nt long and also
include 20 nt PCR adapters (in two groups) on each end.

They were designed using a pull from NCBI's Influenza virus database
on 2018-10-20 for Influenza A, B, and C virus species (see
b12cc141c87e898260ee51a7469a1ac3c890fe93 for details on the datasets
used as input). In the design, each dataset represented a segment
of each virus (8 for Influenza A virus, 8 for Influenza B virus,
and 7 for Influenza C virus), so there were 23 datasets in total.