Add module to filter probes with poly(A) sequence
This adds polya_filter, which is a filter on probes that contain
sequence with long stretches of A bases (or T bases). It might
be desirable to avoid such probe sequences, e.g., to not enrich
carrier RNA. The definition of such a stretch is given as
parameters when initializing a PolyAFilter object.

This also adds unit tests for this module.