Add --small-seq-skip argument to design.py
Sometimes a sequence in an input dataset is too short; in particular,
if it is shorter than the probe length, candidate_probes is unable
to design candidate probes from it and raises an error. One way
to fix this is with the existing --small-seq-min argument, which
allows designing short probes against those sequences.

This argument allows the candidate_probes module to skip over
sequences that are shorter than a specified value (typically
the probe length minus 1). It is possible that such sequences
will not be covered in the design (i.e., if no other candidate
probe covers them).