Fix bug when a sequence is shorter than k-mer length
A bug was introduced by allowing design when sequences are
shorter than the probe length (c09c4f6). If a sequence is
shorter than the k-mer length (k) used for finding probe
coverage in the sequence, then probe.find_probe_covers_in_sequence()
could crash.

In particular, it could compute a bounds_size that is 0 (or
even negative), and then the line that follows:
  for start in range(0, len(sequence) - k + 1, bounds_size):
would fail because the step size is 0 (or lead to undefined
behavior if the end of the range, len(sequence)-k+1, is <0
and bounds_size<0 as well).

This can happen for two reasons: (1) if a sequence is extremely
small (e.g., <20 nt). (2) k is sufficiently high so that
len(sequence) < k. The latter can happen, for example, if
the number of mismatches to tolerate is set to 0 and then
probe._construct_pigeonholed_kmer_probe_map() sets k to be
equal to the probe length, but the input sequence is
shorter than the probe length.

This commit fixes this bug by inserting a check for this
special case, and simply not attempting to cover the sequence
(returning that no probes cover it) if it is true.

This commit also adds a unit test that failed before the fix,
but passes with the fix.