Cluster sequences, if desired, and run filters on each cluster
This commit updates ProbeDesigner to cluster genomes, if desired, and
design probes separately for each cluster. Specifically, it clusters
sequences to create one genome per sequence. Then, for each cluster, it
creates candidate probes and runs the filters that should be run
separately on each cluster. Finally, it merges the designed probes from
each cluster and runs the remaining filters on the combined output
probes.