Pass along target_genomes as a method argument
This commit attempts to make sure that filter objects do not store state
dependent on what they are filtering (probes and their target genomes).
This is necessary to re-use the same filter object multiple times, e.g.,
on each cluster of target genomes (in which there is a different set of
candidate probes and target genomes).

Previously, ProbeDesigner set the target_genomes variable of each filter
object (`f.target_genomes = self.genomes` for a filter `f`) during the
design. This commit changes it to pass along target_genomes through the
filter() method as an argument.

This change necessitated many small changes to unit tests, which are
also a part of this commit.