Include the full ends of sequences when fragmenting genomes
Previously, when fragmenting a sequence in a genome, it would be
possible that the last fragment in the sequence is less than the desired
fragment length (if the sequence length is not a multiple of the
fragment length). This could lead to a short fragment that leads to
errors downstream (for example, too short to construct a MinHash
signature).

This commit ensures that the last fragment in a sequence is of length
equal to the desired fragment length, even if that means two fragments
(near the end of a sequence) overlap.